Add sub-group and location-based membership logic.
   * Move LocationCodes to their own "app"
   * Allow groups to have subgroups and locations
   * Subgroup membership implies membership in parents
   * Member zip or municipality implies membership in location groups
   * Allow location codes that represent groups of other location codes

This is a somewhat complicated change to the models; the purpose of
this is to be able to represent relatively complicated relationships
between groups and locations and calculate dynamic membership lists
based on those rules.

Making this dynamic is an attempt to avoid having to modify the
underlying membership data when implementing some of the complex
rules and relationships being requested by the different Icelandic
pirate orgs. Leaving the source data untouched seems critical as a
matter of principle to be able to respect members' expressed wishes.